fix recipe create bug

+++ backend/foodgram/users/views.py
@@ -17,42 +17,58 @@ User = CustomUser()
 class CustomUserViewSet(UserViewSet):
     pagination_class = PageNumberPagination
 
-    @action(detail=True, methods=['POST'], permission_classes=[IsAuthenticated])
+    @action(detail=True, methods=['POST', 'DELETE'], permission_classes=[IsAuthenticated])
     def subscribe(self, request, id=None):
         user = request.user
         author = get_object_or_404(CustomUser, id=id)
+        if request.method == 'POST':
+            user = request.user
+            author = get_object_or_404(CustomUser, id=id)
 
-        if user == author:
+            if user == author:
+                return Response({
+                    'errors': 'Вы не можете подписываться на самого себя'
+                }, status=status.HTTP_400_BAD_REQUEST)
+            if Subscribe.objects.filter(user=user, author=author).exists():
+                return Response({
+                    'errors': 'Вы уже подписаны на данного пользователя'
+                }, status=status.HTTP_400_BAD_REQUEST)
+
+            follow = Subscribe.objects.create(user=user, author=author)
+            serializer = SubscribeSerializer(
+                follow, context={'request': request}
+            )
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        else:
+            if user == author:
+                return Response({
+                    'errors': 'Вы не можете отписываться от самого себя'
+                }, status=status.HTTP_400_BAD_REQUEST)
+            follow = Subscribe.objects.filter(user=user, author=author)
+            if follow.exists():
+                follow.delete()
+                return Response(status=status.HTTP_204_NO_CONTENT)
+
             return Response({
-                'errors': 'Вы не можете подписываться на самого себя'
-            }, status=status.HTTP_400_BAD_REQUEST)
-        if Subscribe.objects.filter(user=user, author=author).exists():
-            return Response({
-                'errors': 'Вы уже подписаны на данного пользователя'
+                'errors': 'Вы уже отписались'
             }, status=status.HTTP_400_BAD_REQUEST)
 
-        follow = Subscribe.objects.create(user=user, author=author)
-        serializer = SubscribeSerializer(
-            follow, context={'request': request}
-        )
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
-
-    @subscribe.mapping.delete
-    def del_subscribe(self, request, id=None):
-        user = request.user
-        author = get_object_or_404(User, id=id)
-        if user == author:
-            return Response({
-                'errors': 'Вы не можете отписываться от самого себя'
-            }, status=status.HTTP_400_BAD_REQUEST)
-        follow = Subscribe.objects.filter(user=user, author=author)
-        if follow.exists():
-            follow.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
-
-        return Response({
-            'errors': 'Вы уже отписались'
-        }, status=status.HTTP_400_BAD_REQUEST)
+#    @subscribe.mapping.delete
+#    def del_subscribe(self, request, id=None):
+#        user = request.user
+#        author = get_object_or_404(User, id=id)
+#        if user == author:
+#            return Response({
+#                'errors': 'Вы не можете отписываться от самого себя'
+#            }, status=status.HTTP_400_BAD_REQUEST)
+#        follow = Subscribe.objects.filter(user=user, author=author)
+#        if follow.exists():
+#            follow.delete()
+#            return Response(status=status.HTTP_204_NO_CONTENT)
+#
+#        return Response({
+#            'errors': 'Вы уже отписались'
+#        }, status=status.HTTP_400_BAD_REQUEST)
 
     @action(detail=False, methods=['GET'],
             permission_classes=[IsAuthenticated])
